Quality measure
Function compute quality of new image, it is based on number of areas
detected, extension, and image size. This measure is a approach due to
data that know.
Its range is [0,+Inf] where 0 is perfect detection (detect number of
areas equal to number o eyes), and higher is worst, more than some tenth
is bad solution.